Convert the Dates and Times back into Dates and Times, have the database save them and integers (longs)
The primary benefit is that they can be sorted, added, etc. We're treating dates as dates and not as Strings.

They will need to be converted back from longs to dates and times using the Date and Time constructor that takes in longs. Will need to remember to make all dates in GMT/UTC, not in the current timezone, or else they won't load properly.